Compute the context graph in parallel by workers threads.
The workers thread read their work from a randomized input queue to
prevent long trains of accesses to the same backend.
The number of worker threads is the same as the number of nodes in the
cluster to increase concurrency, but not more than some constant.

This modification reduces the elapsed time of computing the context graph
by about 36% in a 3-node cluster running the Guestbook application.

Closes issue #15.