Remove containers, images, processes, minion_status, and version from the API.
Querying minion status will no longer be useful, because minions are going away.

Querying the Cluster Insight version is problematic, because the current,
surprisingly hairy implementation of this is based on talking to the Docker
API, which we will not be doing going forward. It's not clear how essential
this functionality is at this time.

Listing of resources by type is only included in the API because it has been
useful in debugging this software. It's not a feature required by any existing
client. This commit removes listing of processes, containers, and images. In
the case of processes, this is because the collector will not be gathering this
information going forward. In the case of containers and images, it is because
the underlying implementation will be changing radically.