Use built-in type names instead of the types module.
The Python Standard Library documentation says this: "Starting in Python 2.2, built-in factory functions such as int() and str() are also names for the corresponding types. This is now the preferred way to access the type instead of using the types module."

I found the unidiomatic use of the types module distracting when reading this code.