Remove code for determining the project ID.
This code was only able to determine the project ID when the node ID
was of the form "<host name>.c.<project ID>.internal". It's no longer
of that form.

The project ID is a cloud-platform-specific concept. If it is important
to determine the project ID when running Google Cloud Platform in particular,
let's come back to that.